+++ project/plugins.sbt
@@ -3,3 +3,4 @@ addSbtPlugin("org.scala-js"       % "sbt-scalajs"              % "1.3.0")
 addSbtPlugin("com.jsuereth"       % "sbt-pgp"                  % "2.0.1")
 addSbtPlugin("com.typesafe.sbt"   % "sbt-git"                  % "1.0.0")
 addSbtPlugin("org.xerial.sbt"     % "sbt-sonatype"             % "3.9.5")
+addSbtPlugin("ch.epfl.lamp"       % "sbt-dotty"                % "0.4.5")
